# docs/api/OPENAPI.yaml
openapi: 3.1.0
info:
  title: Caladdin API
  version: 1.0.0
  description: |
    HTTP API for Caladdin — AI scheduling assistant.

    **Route groups covered:** `/api/*`, `/book/*`, `/s/*`, `/jobs/*`

    **Authentication:**
    - **Session cookie** (`caladdin_session`) — host-facing routes under `/api/*`
    - **API key** (`x-api-key` header) — cron/worker routes under `/jobs/*`
    - **Guest action token** (signed HMAC, query or body) — cancel/reschedule on `/s/:token/*`
    - **Public** — `/book/*` metadata and `/s/:token` HTML pages (no auth)

    Base URL is `CALADDIN_BASE_URL` (e.g. `https://caladdin.onrender.com`).
servers:
  - url: https://caladdin.onrender.com
    description: Production (example)
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: caladdin_session
      description: HMAC-signed session cookie set by `GET /auth/callback`
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Value of `CALADDIN_API_KEY` env var
  schemas:
...
    WebhookSubscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum: [booking.confirmed, booking.cancelled]
        active:
          type: boolean
        secret:
          type: string
          description: HMAC signing secret (shown once on create)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/webhooks:
    get:
      tags: [Webhooks]
      summary: List webhook subscriptions
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Subscription list
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookSubscription'
        '401':
          description: Not authenticated
    post:
      tags: [Webhooks]
      summary: Create webhook subscription
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum: [booking.confirmed, booking.cancelled]
      responses:
        '201':
          description: Created subscription with secret
        '400':
          description: Invalid URL or events
        '401':
          description: Not authenticated
  /api/webhooks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Webhooks]
      summary: Update webhook subscription
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Updated
        '404':
          description: Not found
    delete:
      tags: [Webhooks]
      summary: Delete webhook subscription
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
...